batcher/requirements.txt
customtkinter
tkinterdnd2
pandas
Pillow